Generate task list documentation file via new cli argument (output-file) #325

+++ src/lib/io.rs
@@ -11,7 +11,7 @@ use rand::distributions::Alphanumeric;
 use rand::{thread_rng, Rng};
 use std::fs::{create_dir_all, remove_file, File};
 use std::io::prelude::*;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::{env, iter};
 
 pub(crate) fn create_text_file(text: &str, extension: &str) -> String {
@@ -74,10 +74,32 @@ pub(crate) fn create_file(write_content: &Fn(&mut File, &str), extension: &str) 
 }
 
 pub(crate) fn delete_file(file: &str) {
-    match remove_file(&file) {
-        Ok(_) => debug!("Temporary file deleted: {}", &file),
-        Err(error) => debug!("Unable to delete temporary file: {} {:#?}", &file, error),
-    };
+    let file_path = Path::new(file);
+
+    if file_path.exists() && file_path.is_file() {
+        match remove_file(&file) {
+            Ok(_) => debug!("File deleted: {}", &file),
+            Err(error) => debug!("Unable to delete file: {} {:#?}", &file, error),
+        };
+    }
+}
+
+pub(crate) fn write_text_file(file_path: &str, text: &str) -> bool {
+    delete_file(&file_path);
+
+    let path = Path::new(file_path);
+    match File::create(&path) {
+        Ok(mut file) => match file.write_all(text.as_bytes()) {
+            Ok(_) => {
+                match file.sync_all() {
+                    Ok(_) => true,
+                    Err(_) => false,
+                }
+            }
+            Err(_) => false,
+        },
+        Err(_) => false,
+    }
 }
 
 pub(crate) fn read_text_file(file_path: &PathBuf) -> String {
